do GPU device memory accounting with doubles
using (signed) int will overflow at 2GB,
switching to unsigned has risk of hiding
overflows and using long long is not as
portable as double precisiong floating point.

+++ lib/gpu/pair_gpu_atom.h
@@ -311,7 +311,7 @@ class PairGPUAtom {
   }
   
   /// Return number of bytes used on device
-  inline int gpu_bytes() { return _gpu_bytes; }
+  inline double gpu_bytes() { return _gpu_bytes; } 
 
   // -------------------------COPY FROM GPU -------------------------------
 
@@ -394,7 +394,7 @@ class PairGPUAtom {
   int *_ilist;
   double _time_cast;
   
-  int _gpu_bytes;
+  double _gpu_bytes;
 
   #ifndef USE_OPENCL
   CUDPPConfiguration sort_config;
